Bug Found in Source Code
I have been working on this for quite some time. I believe that I found a bud in the original source code. It the using guesses a word that has duplicates letters and the letter is in the winning word, then both of the guess letters will turn the color yellow. Regardless if the winning word has the letter in the word once or twice. For example if the winning word is water and I guessed three, then I the letters "t " and "e" would be yellow and the other "e" would be green despite the fact that no other "e's" are in the winning word.