Getting Closing
I'm getting closer to the desired output.  It's hard to explain so I am just going to type my guesses.
Winning word: GNASH
Attempt 1: Audio
Attempt 2: Altar
Attempt 3: Shack
Attempt: 4 Again
I used the word "again" for the last 2 attempts. The first three attempts/ guess showed the desired results. When the word "again" was enter the first "a" was yellow and the second  "a" was green. Ideally only the middle "a" should be green and the reset in their respective colors. Clearly I still have an issue with duplicate letters but it's getting better every so slightly.  Essentially, the issue is that if a letter is present twice and at least one is in the correct position then I get an error.