I Believe This Works!
I made a few changes. It turns out that I needed to decrement the hashmap when the correct words were in the correct position otherwise I run the risk of decrementing the value in the hashmap when the letter is not in the correct position AND when the letter is in the correct position. In the last test I ran is as follows:
Winning word: "FIGHT
Attempt 1: AUDIO
Attempt 2: HEIGH
Attempt 3: FIFTY
Attempt 4: THEFT
Attempt 5: FIGHT
In the first attempt only the letter "i" was yellow as desired. In the second attempt the first "h" was yellow and the second "h" was gray signally that there is only one "h" in the target word. The letter "i" and "g" were yellow as desired. In the third attempt the first "f" was green and the second "f' was grey since there is only one "f" in the target word. The letter "i" was also green and the letter "t" was yellow. In the fourth attempt the last "t" was green and the first "t" was gray. Meaning that there is no other "t" in the word and the "t" in the last position is in the correct spot. The "h" and "f" were yellow and the rest of the letters were gray. Clearly the last attempt was the right word and all of the letter were green. I have a screen shot of my results if anyone would like to see or if they thought my description was confusing. I will be deleting the commented out code as it's just taking up space right now.