Working on Making a Stopwatch
I am currently working on making a stopwatch class. I found a Youtube tutorial that I thought was helping and started adding some of the relevant to the file Stopwatch.java. Also I was getting an error that saying that the libraries that I was trying to import were not accessible. It turns out that the fix to this problem was to just add the line "requires java.desktop;" in the module-info.java class. Not sure why this fixes the issues, but I'm just happy that it works.