I manage to break everything
I managed to break everything. To save myself the hassle I tried to multiple things. I ended up removing and then cloning the repository again. So a lot of the changes in this commit aren't even my own work. I just going to talk about the file that I added. I added three files the JavaFX, Stopwatch, and version. As of now version is completely busted, The other two files work and I an currently trying to implement time trial mode uses a combination of the two working files. I'm not making a lot of progress, but right now a timer appears when the user enter their first word.