I'm trying something out
A lot of the features that I am suppose to implement require code to be added to the method onEnterPressed. I could easily just do this with an extremely long if statement, however I find that to be undesirable. Hence right now I am trying to implement a solution that is similar to how my partner and I handled the playstation lab. We used delegation and used an interface for the rate strategies. This code complies however,  the if statement at lie 66 is always false. Not sure how to go about this, but I'll keep trying.