Changes to updateRowColors
I went back again to change the method updateRowColors. It's fine the way it is. The other small change I made is to the limited guess mode. I have it set so that an attempt is only counted when a 5 letter word is entered.  The previous implementation counted an attempt whenever toast was called.