I fixed a really small aesthetic issue
In call we ran the program and the winning word was "stood".  The we guessed the word "shows". This was to test to make sure that both S's in the word "shows" would not show up yellow. One one "s" was green and the other was yellow. However on the keyboard the "s" was yellow and not green. I believe the added code fixes this. I handled it the same way I did it for the updateRowColor method. I tried this code and the winning word was "vital" I tried the word vivid and  the "v" and "i" were green and the "d" was dark grey as desired.