Added a label so that the user knows what game mode they are in
I added a label so that the users know what game mode they are playing the game. I did the by adding a vertical box. This vertical box contains a label for the game mode and the number of attempts for the limited guesses game mode. Lastly I played around with some of the values in the file MainApplication and main-view.fxml to see if I could get the output to look a little better (meaning more aligned and or centered)